[#47] refactor : 컴포넌트 폴더링 변경
