+++ components/Healthlog/QuickButtons/index.tsx
